requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
babel==2.16.0
billiard==4.2.1
celery==5.4.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2024.10
cron-descriptor==1.4.5
decorator==5.1.1
diff-match-patch==20230430
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==65.0.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-import-export==4.1.1
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-scheduler==0.10.1
django-slick-reporting==1.3.1
django-timezone-field==7.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
flower==2.0.1
humanize==4.10.0
icalendar==5.0.13
ipython==8.18.1
jedi==0.19.1
kombu==5.4.2
matplotlib-inline==0.1.7
numpy==2.0.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
phonenumberslite==8.13.47
pillow==11.0.0
pkg_resources==0.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.0.8
reportlab==4.2.5
schedule==1.2.2
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tablib==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
u==0.1
vine==5.1.0
wcwidth==0.2.13
